Refactor runner and utils
